Split extractors/skills into models/adapters/extractor, carry refusal text
Per review feedback on the monolithic extractors/skills.py, split it along the seam
the per-harness branching already followed:

- models.py: what the extractors return (AvailableSkill, InvokedSkill)
- adapters.py: the per-harness block shapes. The tool-call and tool-result elif
  chains become one named recognizer per harness, tried in the same order as
  before, returning a common ToolCallBlock / ToolResultBlock.
- extractor.py: the harness-independent decisions (which calls are skill loads,
  pairing calls with results, deduplication) and the public entry points.
- _patterns.py / _normalize.py: the harness literals and the shared primitives.

Public API is unchanged: strands_evals.extractors still exports the same four
names. Verified by dumping extraction output over every fixture before and after;
the only difference is the new field below.

InvokedSkill gains error: str | None, the harness's refusal message on a failed
load. The text was recognized and then discarded, so "Skill 'pdf-procesing' not
found. Available skills: pdf-processing" (a misspelled name) and "Available
skills: (none)" (a harness that mounted nothing) both collapsed into "the load
failed". It now reaches the adherence reason, the SkillInvoked reason, and the
selection judge's prompt, where a name the harness did not recognize is a worse
pick than a right one it could not mount.

serialize_trajectory moves to evaluators/prompt_templates, since it is prompt
formatting rather than extraction and it is new public API in this branch, so
moving it later would be a break.
